refactor: 패키지 경로 정리 (팀 규칙 적용)
